requirements.txt
absl-py==1.0.0
astunparse==1.6.3
cachetools==3.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==6.7
cloudpickle
cycler
cytoolz==0.11.0
dask==2021.10.0
Flask==1.1.1
flask-cors
flatbuffers==2.0
fonttools==4.25.0
fsspec
futures==3.1.1
gast==0.4.0
google-assistant-grpc==0.2.1
google-assistant-sdk==0.5.1
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.5.5
grpcio==1.43.0
gunicorn==19.9.0
h5py==2.9.0
idna==3.3
imagecodecs
imageio
importlib-metadata==4.10.0
ipaddress==1.0.22
itsdangerous==1.1.0
Jinja2==2.10.1
keras==2.7.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver
libclang==12.0.0
locket==0.2.1
Markdown==3.3.6
MarkupSafe==1.1.1
matplotlib
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
munkres==1.1.4
networkx
numpy
oauthlib==2.1.0
olefile==0.46
opt-einsum==3.3.0
packaging
pandas==1.3.5
partd
pathlib2==2.3.3
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycparser==2.21
pyparsing
python-dateutil
pytz==2021.3
PyWavelets
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.0.0
rsa==4.0
scikit-image==0.18.3
scikit-learn==0.20.2
scipy
six
sounddevice==0.3.12
tenacity==4.12.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
tifffile
toolz
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==0.15.5
wincertstore==0.2
wrapt==1.13.3
WSGIserver==1.3
zipp==3.6.0
pytest==6.2.2
-e .
